// dating-app/src/app/models/member.model.ts
import { Photo } from "./photo.model";

export interface Member {
    id: number;
    userName: string;
    photoUrl: string;
    age: number;
    knowAs: string;
    created: Date;
    lastActive: Date;
    gender: string;
    introduction: string;
    lookingFor: string;
    interests: string;
    city: string;
    country: string;
    photos: Photo[];
}